v1.5.0 #19 Fix for OSX random custom color actions Scroll to cursor on markerlist select Snapshot paste into empty project creates new tracks by default Actions for show all tracks in MCP/TCP Move cursor to peak amplitude in selected item Select prev/next track doesn't unselect at extents Scroll left/right by 10% and 1%

+++ Snapshots/Snapshots.cpp
@@ -177,7 +177,7 @@ void ImportSnapshot()
 			int iLen = 0;
 			while(!cfg->GetLine(str, 4096))
 			{
-				int iNewLen = iLen + strlen(str);
+				int iNewLen = iLen + (int)strlen(str);
 				chunk.Resize(iNewLen + 2);
 				strcpy(chunk.Get()+iLen, str);
 				strcpy(chunk.Get()+iNewLen, "\n");
